# optimized add_weighted