Reduced the number of temporary variables to optimize the weighted sum


Device: cpu

=== Shape: (8, 3, 64, 64) ===
Scalar params -> max abs diff = 0.000e+00
Per-element params -> max abs diff = 0.000e+00
Scalar params: Original: 0.103 ms | Optimized: 0.080 ms | Speedup: 1.285x
Per-element:    Original: 0.104 ms | Optimized: 0.090 ms | Speedup: 1.145x

=== Shape: (16, 3, 514, 514) ===
Scalar params -> max abs diff = 0.000e+00
Per-element params -> max abs diff = 0.000e+00
Scalar params: Original: 77.331 ms | Optimized: 44.450 ms | Speedup: 1.740x
Per-element:    Original: 83.215 ms | Optimized: 51.482 ms | Speedup: 1.616x

https://colab.research.google.com/drive/1DzmpICYS5QJd09ZMMiKecFN65DIv_DMp?usp=sharing